# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shaghlny - Job Hiring Platform API",
        "description": "API documentation for Shaghlny - Job Hiring Platform",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/jobs": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get list of jobs with recommendations",
                "description": "Fetch jobs and recommended jobs based on user's profile skills.",
                "operationId": "9538d30d530d4c3680ea31fe228944e2",
                "responses": {
                    "200": {
                        "description": "List of jobs with recommendations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "jobs": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Job"
                                            }
                                        },
                                        "recommended_jobs": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Job"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Post a new job",
                "description": "Create a new job listing.",
                "operationId": "36deb2a6480382cf8385911c10aeebf0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "job_type"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Software Developer"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Job description here..."
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Remote"
                                    },
                                    "job_type": {
                                        "type": "string",
                                        "enum": [
                                            "Full-time",
                                            "Part-time",
                                            "Contract"
                                        ]
                                    },
                                    "salary": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Job posted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Job posted successfully"
                                        },
                                        "job": {
                                            "$ref": "#/components/schemas/Job"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/company/jobs": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get jobs with questions",
                "description": "Fetch jobs that have associated questions.",
                "operationId": "014f49865ad5ee38bb99589d1765cc0d",
                "responses": {
                    "200": {
                        "description": "Jobs with questions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "jobs_with_questions": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Job"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jobs/applied": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get applied jobs",
                "description": "Fetch jobs that the authenticated user has applied to.",
                "operationId": "07994410966b36ce285af27f2d2a20c6",
                "responses": {
                    "200": {
                        "description": "List of applied jobs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "applied_jobs": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Job"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jobs/{jobId}/apply": {
            "post": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Apply for a job",
                "description": "Apply for a job if not already applied.",
                "operationId": "860dfc3e627b26de2d4031333613c94d",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully applied for the job",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully applied for the job"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Already applied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Already applied"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Job not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jobs/{jobId}": {
            "put": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Update a job",
                "description": "Update an existing job listing.",
                "operationId": "07a4daa8c53e9c111508028f937ae498",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Job Title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated job description"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "New Location"
                                    },
                                    "salary": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 60000
                                    },
                                    "job_type": {
                                        "type": "string",
                                        "enum": [
                                            "Full-time",
                                            "Part-time",
                                            "Contract"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Job updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Job updated successfully"
                                        },
                                        "job": {
                                            "$ref": "#/components/schemas/Job"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Job not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Delete a job",
                "description": "Delete an existing job listing.",
                "operationId": "769a235b882a64edbccbc9bbb06b4806",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Job deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Job not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jobs/recommended": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get recommended jobs based on user skills",
                "description": "Get jobs that match the user's profile skills.",
                "operationId": "abd6ac49f213fe7fc0f3a55d210cde32",
                "responses": {
                    "200": {
                        "description": "List of recommended jobs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "recommended_jobs": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Job"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Job": {
                "required": [
                    "title",
                    "description",
                    "job_type"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "title": {
                        "type": "string",
                        "example": "Software Developer"
                    },
                    "description": {
                        "type": "string",
                        "example": "Job description here..."
                    },
                    "location": {
                        "type": "string",
                        "example": "Remote"
                    },
                    "salary": {
                        "type": "number",
                        "format": "float",
                        "example": 50000
                    },
                    "job_type": {
                        "type": "string",
                        "enum": [
                            "Full-time",
                            "Part-time",
                            "Contract"
                        ],
                        "example": "Full-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-25T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-25T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}